+++ src/main/java/com/alpha67/amc/procedures/FusionMachinProcedureProcedure.java
@@ -54,7 +54,7 @@ public class FusionMachinProcedureProcedure {
 					return tileEntity.getTileData().getDouble(tag);
 				return -1;
 			}
-		}.getValue(world, new BlockPos((int) x, (int) y, (int) z), "recipe"));
+		}.getValue(world, new BlockPos(x, y, z), "recipe"));
 		if (((new Object() {
 			public ItemStack getItemStack(BlockPos pos, int sltid) {
 				AtomicReference<ItemStack> _retval = new AtomicReference<>(ItemStack.EMPTY);
@@ -66,7 +66,7 @@ public class FusionMachinProcedureProcedure {
 				}
 				return _retval.get();
 			}
-		}.getItemStack(new BlockPos((int) x, (int) y, (int) z), (int) (1))).getItem() == RefinedRubyItem.block && (new Object() {
+		}.getItemStack(new BlockPos(x, y, z), (int) (1))).getItem() == RefinedRubyItem.block && (new Object() {
 			public ItemStack getItemStack(BlockPos pos, int sltid) {
 				AtomicReference<ItemStack> _retval = new AtomicReference<>(ItemStack.EMPTY);
 				TileEntity _ent = world.getTileEntity(pos);
@@ -77,7 +77,7 @@ public class FusionMachinProcedureProcedure {
 				}
 				return _retval.get();
 			}
-		}.getItemStack(new BlockPos((int) x, (int) y, (int) z), (int) (2))).getItem() == Items.GOLD_INGOT && (new Object() {
+		}.getItemStack(new BlockPos(x, y, z), (int) (2))).getItem() == Items.GOLD_INGOT && (new Object() {
 			public ItemStack getItemStack(BlockPos pos, int sltid) {
 				AtomicReference<ItemStack> _retval = new AtomicReference<>(ItemStack.EMPTY);
 				TileEntity _ent = world.getTileEntity(pos);
@@ -88,7 +88,7 @@ public class FusionMachinProcedureProcedure {
 				}
 				return _retval.get();
 			}
-		}.getItemStack(new BlockPos((int) x, (int) y, (int) z), (int) (0))).getItem() == Items.LAVA_BUCKET || (new Object() {
+		}.getItemStack(new BlockPos(x, y, z), (int) (0))).getItem() == Items.LAVA_BUCKET || (new Object() {
 			public ItemStack getItemStack(BlockPos pos, int sltid) {
 				AtomicReference<ItemStack> _retval = new AtomicReference<>(ItemStack.EMPTY);
 				TileEntity _ent = world.getTileEntity(pos);
@@ -99,7 +99,7 @@ public class FusionMachinProcedureProcedure {
 				}
 				return _retval.get();
 			}
-		}.getItemStack(new BlockPos((int) x, (int) y, (int) z), (int) (2))).getItem() == RefinedRubyItem.block && (new Object() {
+		}.getItemStack(new BlockPos(x, y, z), (int) (2))).getItem() == RefinedRubyItem.block && (new Object() {
 			public ItemStack getItemStack(BlockPos pos, int sltid) {
 				AtomicReference<ItemStack> _retval = new AtomicReference<>(ItemStack.EMPTY);
 				TileEntity _ent = world.getTileEntity(pos);
@@ -110,7 +110,7 @@ public class FusionMachinProcedureProcedure {
 				}
 				return _retval.get();
 			}
-		}.getItemStack(new BlockPos((int) x, (int) y, (int) z), (int) (1))).getItem() == Items.GOLD_INGOT && (new Object() {
+		}.getItemStack(new BlockPos(x, y, z), (int) (1))).getItem() == Items.GOLD_INGOT && (new Object() {
 			public ItemStack getItemStack(BlockPos pos, int sltid) {
 				AtomicReference<ItemStack> _retval = new AtomicReference<>(ItemStack.EMPTY);
 				TileEntity _ent = world.getTileEntity(pos);
@@ -121,7 +121,7 @@ public class FusionMachinProcedureProcedure {
 				}
 				return _retval.get();
 			}
-		}.getItemStack(new BlockPos((int) x, (int) y, (int) z), (int) (0))).getItem() == Items.LAVA_BUCKET) && ((new Object() {
+		}.getItemStack(new BlockPos(x, y, z), (int) (0))).getItem() == Items.LAVA_BUCKET) && ((new Object() {
 			public ItemStack getItemStack(BlockPos pos, int sltid) {
 				AtomicReference<ItemStack> _retval = new AtomicReference<>(ItemStack.EMPTY);
 				TileEntity _ent = world.getTileEntity(pos);
@@ -132,7 +132,7 @@ public class FusionMachinProcedureProcedure {
 				}
 				return _retval.get();
 			}
-		}.getItemStack(new BlockPos((int) x, (int) y, (int) z), (int) (3))).getItem() == MetalRubyIngotItem.block && new Object() {
+		}.getItemStack(new BlockPos(x, y, z), (int) (3))).getItem() == MetalRubyIngotItem.block && new Object() {
 			public int getAmount(IWorld world, BlockPos pos, int sltid) {
 				AtomicInteger _retval = new AtomicInteger(0);
 				TileEntity _ent = world.getTileEntity(pos);
@@ -143,7 +143,7 @@ public class FusionMachinProcedureProcedure {
 				}
 				return _retval.get();
 			}
-		}.getAmount(world, new BlockPos((int) x, (int) y, (int) z), (int) (3)) <= 64 || new Object() {
+		}.getAmount(world, new BlockPos(x, y, z), (int) (3)) <= 64 || new Object() {
 			public int getAmount(IWorld world, BlockPos pos, int sltid) {
 				AtomicInteger _retval = new AtomicInteger(0);
 				TileEntity _ent = world.getTileEntity(pos);
@@ -154,9 +154,9 @@ public class FusionMachinProcedureProcedure {
 				}
 				return _retval.get();
 			}
-		}.getAmount(world, new BlockPos((int) x, (int) y, (int) z), (int) (3)) == 0)) {
+		}.getAmount(world, new BlockPos(x, y, z), (int) (3)) == 0)) {
 			if (!world.isRemote()) {
-				BlockPos _bp = new BlockPos((int) x, (int) y, (int) z);
+				BlockPos _bp = new BlockPos(x, y, z);
 				TileEntity _tileEntity = world.getTileEntity(_bp);
 				BlockState _bs = world.getBlockState(_bp);
 				if (_tileEntity != null)
@@ -171,10 +171,10 @@ public class FusionMachinProcedureProcedure {
 						return tileEntity.getTileData().getDouble(tag);
 					return -1;
 				}
-			}.getValue(world, new BlockPos((int) x, (int) y, (int) z), "timer")));
+			}.getValue(world, new BlockPos(x, y, z), "timer")));
 		} else {
 			if (!world.isRemote()) {
-				BlockPos _bp = new BlockPos((int) x, (int) y, (int) z);
+				BlockPos _bp = new BlockPos(x, y, z);
 				TileEntity _tileEntity = world.getTileEntity(_bp);
 				BlockState _bs = world.getBlockState(_bp);
 				if (_tileEntity != null)
@@ -183,7 +183,7 @@ public class FusionMachinProcedureProcedure {
 					((World) world).notifyBlockUpdate(_bp, _bs, _bs, 3);
 			}
 			if (!world.isRemote()) {
-				BlockPos _bp = new BlockPos((int) x, (int) y, (int) z);
+				BlockPos _bp = new BlockPos(x, y, z);
 				TileEntity _tileEntity = world.getTileEntity(_bp);
 				BlockState _bs = world.getBlockState(_bp);
 				if (_tileEntity != null)
@@ -199,9 +199,9 @@ public class FusionMachinProcedureProcedure {
 					return tileEntity.getTileData().getDouble(tag);
 				return -1;
 			}
-		}.getValue(world, new BlockPos((int) x, (int) y, (int) z), "recipe")) {
+		}.getValue(world, new BlockPos(x, y, z), "recipe")) {
 			if (!world.isRemote()) {
-				BlockPos _bp = new BlockPos((int) x, (int) y, (int) z);
+				BlockPos _bp = new BlockPos(x, y, z);
 				TileEntity _tileEntity = world.getTileEntity(_bp);
 				BlockState _bs = world.getBlockState(_bp);
 				if (_tileEntity != null)
@@ -210,7 +210,7 @@ public class FusionMachinProcedureProcedure {
 					((World) world).notifyBlockUpdate(_bp, _bs, _bs, 3);
 			}
 			if (!world.isRemote()) {
-				BlockPos _bp = new BlockPos((int) x, (int) y, (int) z);
+				BlockPos _bp = new BlockPos(x, y, z);
 				TileEntity _tileEntity = world.getTileEntity(_bp);
 				BlockState _bs = world.getBlockState(_bp);
 				if (_tileEntity != null)
@@ -225,7 +225,7 @@ public class FusionMachinProcedureProcedure {
 						return tileEntity.getTileData().getDouble(tag);
 					return -1;
 				}
-			}.getValue(world, new BlockPos((int) x, (int) y, (int) z), "recipe") == 1) {
+			}.getValue(world, new BlockPos(x, y, z), "recipe") == 1) {
 				if (new Object() {
 					public double getValue(IWorld world, BlockPos pos, String tag) {
 						TileEntity tileEntity = world.getTileEntity(pos);
@@ -233,7 +233,7 @@ public class FusionMachinProcedureProcedure {
 							return tileEntity.getTileData().getDouble(tag);
 						return -1;
 					}
-				}.getValue(world, new BlockPos((int) x, (int) y, (int) z), "timer") >= 100) {
+				}.getValue(world, new BlockPos(x, y, z), "timer") >= 100) {
 					if (new Object() {
 						public double getValue(IWorld world, BlockPos pos, String tag) {
 							TileEntity tileEntity = world.getTileEntity(pos);
@@ -241,7 +241,7 @@ public class FusionMachinProcedureProcedure {
 								return tileEntity.getTileData().getDouble(tag);
 							return -1;
 						}
-					}.getValue(world, new BlockPos((int) x, (int) y, (int) z), "recipe") == 1) {
+					}.getValue(world, new BlockPos(x, y, z), "recipe") == 1) {
 						if (new Object() {
 							public double getValue(IWorld world, BlockPos pos, String tag) {
 								TileEntity tileEntity = world.getTileEntity(pos);
@@ -249,9 +249,9 @@ public class FusionMachinProcedureProcedure {
 									return tileEntity.getTileData().getDouble(tag);
 								return -1;
 							}
-						}.getValue(world, new BlockPos((int) x, (int) y, (int) z), "bucketoflava") >= 1) {
+						}.getValue(world, new BlockPos(x, y, z), "bucketoflava") >= 1) {
 							{
-								TileEntity _ent = world.getTileEntity(new BlockPos((int) x, (int) y, (int) z));
+								TileEntity _ent = world.getTileEntity(new BlockPos(x, y, z));
 								if (_ent != null) {
 									final int _sltid = (int) (0);
 									final int _amount = (int) 1;
@@ -265,7 +265,7 @@ public class FusionMachinProcedureProcedure {
 								}
 							}
 							{
-								TileEntity _ent = world.getTileEntity(new BlockPos((int) x, (int) y, (int) z));
+								TileEntity _ent = world.getTileEntity(new BlockPos(x, y, z));
 								if (_ent != null) {
 									final int _sltid = (int) (1);
 									final int _amount = (int) 1;
@@ -279,7 +279,7 @@ public class FusionMachinProcedureProcedure {
 								}
 							}
 							{
-								TileEntity _ent = world.getTileEntity(new BlockPos((int) x, (int) y, (int) z));
+								TileEntity _ent = world.getTileEntity(new BlockPos(x, y, z));
 								if (_ent != null) {
 									final int _sltid = (int) (2);
 									final int _amount = (int) 1;
@@ -293,7 +293,7 @@ public class FusionMachinProcedureProcedure {
 								}
 							}
 							{
-								TileEntity _ent = world.getTileEntity(new BlockPos((int) x, (int) y, (int) z));
+								TileEntity _ent = world.getTileEntity(new BlockPos(x, y, z));
 								if (_ent != null) {
 									final int _sltid = (int) (3);
 									final ItemStack _setstack = new ItemStack(MetalRubyIngotItem.block);
@@ -308,7 +308,7 @@ public class FusionMachinProcedureProcedure {
 											}
 											return _retval.get();
 										}
-									}.getAmount(world, new BlockPos((int) x, (int) y, (int) z), (int) (3)) + 1));
+									}.getAmount(world, new BlockPos(x, y, z), (int) (3)) + 1));
 									_ent.getCapability(CapabilityItemHandler.ITEM_HANDLER_CAPABILITY, null).ifPresent(capability -> {
 										if (capability instanceof IItemHandlerModifiable) {
 											((IItemHandlerModifiable) capability).setStackInSlot(_sltid, _setstack);
@@ -317,7 +317,7 @@ public class FusionMachinProcedureProcedure {
 								}
 							}
 							if (!world.isRemote()) {
-								BlockPos _bp = new BlockPos((int) x, (int) y, (int) z);
+								BlockPos _bp = new BlockPos(x, y, z);
 								TileEntity _tileEntity = world.getTileEntity(_bp);
 								BlockState _bs = world.getBlockState(_bp);
 								if (_tileEntity != null)
@@ -328,14 +328,14 @@ public class FusionMachinProcedureProcedure {
 												return tileEntity.getTileData().getDouble(tag);
 											return -1;
 										}
-									}.getValue(world, new BlockPos((int) x, (int) y, (int) z), "bucketoflava")) - 1));
+									}.getValue(world, new BlockPos(x, y, z), "bucketoflava")) - 1));
 								if (world instanceof World)
 									((World) world).notifyBlockUpdate(_bp, _bs, _bs, 3);
 							}
 						}
 					}
 					if (!world.isRemote()) {
-						BlockPos _bp = new BlockPos((int) x, (int) y, (int) z);
+						BlockPos _bp = new BlockPos(x, y, z);
 						TileEntity _tileEntity = world.getTileEntity(_bp);
 						BlockState _bs = world.getBlockState(_bp);
 						if (_tileEntity != null)
@@ -345,7 +345,7 @@ public class FusionMachinProcedureProcedure {
 					}
 				}
 				if (!world.isRemote()) {
-					BlockPos _bp = new BlockPos((int) x, (int) y, (int) z);
+					BlockPos _bp = new BlockPos(x, y, z);
 					TileEntity _tileEntity = world.getTileEntity(_bp);
 					BlockState _bs = world.getBlockState(_bp);
 					if (_tileEntity != null)
@@ -356,7 +356,7 @@ public class FusionMachinProcedureProcedure {
 									return tileEntity.getTileData().getDouble(tag);
 								return -1;
 							}
-						}.getValue(world, new BlockPos((int) x, (int) y, (int) z), "timer") + 0));
+						}.getValue(world, new BlockPos(x, y, z), "timer") + 0));
 					if (world instanceof World)
 						((World) world).notifyBlockUpdate(_bp, _bs, _bs, 3);
 				}
