+++ src/main/java/com/alpha67/amc/procedures/Lavabuketprocedure0Procedure.java
@@ -42,7 +42,7 @@ public class Lavabuketprocedure0Procedure {
 					return tileEntity.getTileData().getDouble(tag);
 				return -1;
 			}
-		}.getValue(world, new BlockPos((int) x, (int) y, (int) z), "bucketoflava") == 0) {
+		}.getValue(world, new BlockPos(x, y, z), "bucketoflava") == 0) {
 			return true;
 		}
 		return false;
