+++ src/main/java/com/alpha67/amc/procedures/Lavabuketprocedure4Procedure.java
@@ -42,7 +42,7 @@ public class Lavabuketprocedure4Procedure {
 					return tileEntity.getTileData().getDouble(tag);
 				return -1;
 			}
-		}.getValue(world, new BlockPos((int) x, (int) y, (int) z), "bucketoflava") == 4) {
+		}.getValue(world, new BlockPos(x, y, z), "bucketoflava") == 4) {
 			return true;
 		}
 		return false;
