+++ src/main/java/com/alpha67/amc/procedures/Lavabuketprocedure9Procedure.java
@@ -42,7 +42,7 @@ public class Lavabuketprocedure9Procedure {
 					return tileEntity.getTileData().getDouble(tag);
 				return -1;
 			}
-		}.getValue(world, new BlockPos((int) x, (int) y, (int) z), "bucketoflava") == 9) {
+		}.getValue(world, new BlockPos(x, y, z), "bucketoflava") == 9) {
 			return true;
 		}
 		return false;
